퇴근일지 이미지 관련 예외처리, 트랜잭션 적용
